release 3.1.173

+++ tdmysql/include/tencentcloud/tdmysql/v20211122/model/ServerlessCcu.h
@@ -51,14 +51,14 @@ namespace TencentCloud
                      * @return MinCcu <p>ccu minimum value</p>
                      * 
                      */
-                    int64_t GetMinCcu() const;
+                    double GetMinCcu() const;
 
                     /**
                      * 设置<p>ccu minimum value</p>
                      * @param _minCcu <p>ccu minimum value</p>
                      * 
                      */
-                    void SetMinCcu(const int64_t& _minCcu);
+                    void SetMinCcu(const double& _minCcu);
 
                     /**
                      * 判断参数 MinCcu 是否已赋值
@@ -72,14 +72,14 @@ namespace TencentCloud
                      * @return MaxCcu <p>Maximum value of ccu</p>
                      * 
                      */
-                    std::vector<int64_t> GetMaxCcu() const;
+                    std::vector<double> GetMaxCcu() const;
 
                     /**
                      * 设置<p>Maximum value of ccu</p>
                      * @param _maxCcu <p>Maximum value of ccu</p>
                      * 
                      */
-                    void SetMaxCcu(const std::vector<int64_t>& _maxCcu);
+                    void SetMaxCcu(const std::vector<double>& _maxCcu);
 
                     /**
                      * 判断参数 MaxCcu 是否已赋值
@@ -93,13 +93,13 @@ namespace TencentCloud
                     /**
                      * <p>ccu minimum value</p>
                      */
-                    int64_t m_minCcu;
+                    double m_minCcu;
                     bool m_minCcuHasBeenSet;
 
                     /**
                      * <p>Maximum value of ccu</p>
                      */
-                    std::vector<int64_t> m_maxCcu;
+                    std::vector<double> m_maxCcu;
                     bool m_maxCcuHasBeenSet;
 
                 };
